requirements.txt
accelerate==1.11.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
deep-translator==1.11.4
docx2txt==0.9
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
language_tool_python==2.9.4
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
psutil==7.1.0
PyMuPDF==1.26.4
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tkinterdnd2==0.4.3
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
